Extract /games/ endpoint into app/views/games.py
Moves endpoint_games(), get_game_ids(), get_rosters_for_games(),
build_linescore_and_scoring_plays(), and their helpers out of
stat_retrieval.py into a dedicated file. stat_retrieval.py imports
get_game_ids() for the /events/ and /stats/ endpoints that also use it.

No logic changes — pure file extraction.
